requirements.txt
africastalking==1.2.5
amqp==5.1.1
arabic-reshaper==3.0.0
asgiref==3.7.1
asn1crypto==1.5.1
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
contextlib2==21.6.0
cron-descriptor==1.4.0
cryptography==40.0.2
cssselect2==0.7.0
Django==4.2.1
django-appconf==1.0.5
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-ckeditor==6.5.1
django-cors-headers==4.0.0
django-imagekit==4.1.0
django-js-asset==2.0.0
django-timezone-field==5.0
html5lib==1.1
idna==3.4
kombu==5.2.4
lxml==4.9.2
oscrypto==1.3.0
pilkit==2.0
Pillow==9.5.0
prompt-toolkit==3.0.38
pycparser==2.21
pyHanko==0.18.1
pyhanko-certvalidator==0.22.0
pypdf==3.9.0
pypng==0.20220715.0
python-bidi==0.4.2
python-crontab==2.7.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
PyYAML==6.0
qrcode==7.4.2
redis==4.5.5
reportlab==3.6.13
requests==2.31.0
schema==0.7.5
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.6.2
tzdata==2023.3
tzlocal==5.0.1
uritools==4.0.1
urllib3==2.0.2
vine==5.0.0
wcwidth==0.2.6
webencodings==0.5.1
whitenoise==6.4.0
xhtml2pdf==0.2.11
